requirements.txt
fastapi
uvicorn
pdfplumber
requests
sentence-transformers
faiss-cpu
transformers